+++ .github/workflows/main.yml
@@ -65,4 +65,28 @@ jobs:
           pip install poetry && \
           poetry install
       - name: Run bandit
-        run: poetry run bandit -r . -x ./test
+        run: poetry run bandit -r .
+
+  selenium:
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Check out repository
+        uses: actions/checkout@v4
+
+      - name: Setup Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: 3.11
+
+      - name: Install dependencies
+        run: |
+          sudo apt update -y && \
+          sudo apt install -y sqlcipher libsqlcipher-dev firefox && \
+          python -m pip install --upgrade pip && \
+          pip install poetry && \
+          poetry install
+
+      - name: Run Selenium tests with Poetry
+        run: |
+          poetry run pytest test/ui/ui.py
